update some snapshots and move all the utility test functions to another module
